* origin/pr/19488:
  Use %zu for printf-style formatting of size_t values.

+++ src/advanced_inv.cpp
@@ -352,7 +352,7 @@ void advanced_inventory::print_items( advanced_inventory_pane &pane, bool active
 
         //print weight column
         double it_weight = convert_weight( sitem.weight );
-        size_t w_precision;
+        int w_precision;
         print_color = ( it_weight > 0 ) ? thiscolor : thiscolordark;
 
         if( it_weight >= 1000.0 ) {
